refactor(views): share the latency probe between Outbounds and Endpoints
Endpoints register as outbounds in the core, so both pages probe the same
api/checkOutbound. The whole flow was duplicated: the CheckResult interface,
checkResults, the probe, the test-all wrapper and an eleven-line delayRow that
was byte-identical between the two files.

useDelayCheck keeps the caller-facing names via destructuring, so neither
template changes. It takes a getter rather than a Ref: both callers hold a
readonly computed, and a Ref parameter would type-check and then silently fail
to track.

The list is typed any[] rather than {tag: string}[]. Outbound and Endpoint each
resolve to `{ type: string; [k: string]: any }` — types/outbounds.ts builds its
InterfaceMap as a mapped type that never references the per-protocol interfaces
below it, so OutboundBasics.tag is not part of the exported union and a
required-tag bound does not hold.
