+++ frontend/src/views/Endpoints.vue
@@ -80,7 +80,7 @@
 import { computed, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
 import Data from '@/store/modules/data'
-import HttpUtils from '@/plugins/httputil'
+import { useDelayCheck } from '@/plugins/useDelayCheck'
 import { Endpoint } from '@/types/endpoints'
 import Btn from '@/components/ui/Btn.vue'
 import Ico from '@/components/ui/Ico.vue'
@@ -108,38 +108,13 @@ const onlines = computed(() => [
 // ---------------- delay check ----------------
 // Endpoints register as outbounds in the core, so the same api/checkOutbound
 // latency probe used on the Outbounds page works here unchanged.
-interface CheckResult {
-  loading?: boolean
-  success: boolean
-  data?: { OK?: boolean; Delay?: number; Error?: string } | null
-  errorMessage?: string
-}
-
-const checkResults = ref<Record<string, CheckResult>>({})
-
-const checkEndpoint = async (tag: string) => {
-  checkResults.value = { ...checkResults.value, [tag]: { loading: true, success: false } }
-  const msg = await HttpUtils.get('api/checkOutbound', { tag })
-  const success = msg.success && msg.obj?.OK
-  const errorMessage = success ? undefined : (msg.obj?.Error ?? msg.msg ?? '')
-  checkResults.value = {
-    ...checkResults.value,
-    [tag]: { loading: false, success, data: msg.obj ?? null, errorMessage },
-  }
-}
-
-const testingAll = ref(false)
-
-const checkAllEndpoints = async () => {
-  const list = endpoints.value
-  if (list.length === 0) return
-  testingAll.value = true
-  try {
-    await Promise.all(list.map((e) => checkEndpoint(e.tag)))
-  } finally {
-    testingAll.value = false
-  }
-}
+const {
+  checkResults,
+  testingAll,
+  check: checkEndpoint,
+  checkAll: checkAllEndpoints,
+  delayRow,
+} = useDelayCheck(() => endpoints.value)
 
 // ---------------- card rows ----------------
 const cardRows = (item: any): EntityRow[] => [
@@ -154,20 +129,8 @@ const cardRows = (item: any): EntityRow[] => [
     mono: item.listen_port > 0,
   },
   { k: t('types.wg.peers'), v: item.peers?.length ?? t('ui.none') },
-  delayRow(item),
+  delayRow(item.tag),
 ]
-
-const delayRow = (item: any): EntityRow => {
-  const r = checkResults.value[item.tag]
-  if (r?.loading) return { k: t('out.delay'), v: '…', mono: true }
-  if (r && r.loading == false) {
-    if (r.success) {
-      return { k: t('out.delay'), v: (r.data?.Delay ?? 0) + ' ' + t('date.ms'), mono: true, color: 'var(--emerald)' }
-    }
-    return { k: t('out.delay'), v: r.errorMessage || t('failed'), color: 'var(--rose)' }
-  }
-  return { k: t('out.delay'), v: t('ui.none') }
-}
 
 // ---------------- drawers / modals ----------------
 const drawer = ref({ visible: false, id: 0, data: '' })
