+++ frontend/src/views/Outbounds.vue
@@ -81,8 +81,8 @@
 import { computed, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
 import Data from '@/store/modules/data'
-import HttpUtils from '@/plugins/httputil'
 import { outColor } from '@/plugins/colors'
+import { useDelayCheck } from '@/plugins/useDelayCheck'
 import { Outbound } from '@/types/outbounds'
 import Btn from '@/components/ui/Btn.vue'
 import Ico from '@/components/ui/Ico.vue'
@@ -104,39 +104,14 @@ const outboundTags = computed((): string[] => dataStore.outboundTags)
 
 const onlines = computed(() => dataStore.onlines.outbound ?? [])
 
-// ---------------- delay check (legacy logic) ----------------
-interface CheckResult {
-  loading?: boolean
-  success: boolean
-  data?: { OK?: boolean; Delay?: number; Error?: string } | null
-  errorMessage?: string
-}
-
-const checkResults = ref<Record<string, CheckResult>>({})
-
-const checkOutbound = async (tag: string) => {
-  checkResults.value = { ...checkResults.value, [tag]: { loading: true, success: false } }
-  const msg = await HttpUtils.get('api/checkOutbound', { tag })
-  const success = msg.success && msg.obj?.OK
-  const errorMessage = success ? undefined : (msg.obj?.Error ?? msg.msg ?? '')
-  checkResults.value = {
-    ...checkResults.value,
-    [tag]: { loading: false, success, data: msg.obj ?? null, errorMessage },
-  }
-}
-
-const testingAll = ref(false)
-
-const checkAllOutbounds = async () => {
-  const list = outbounds.value
-  if (list.length === 0) return
-  testingAll.value = true
-  try {
-    await Promise.all(list.map((o) => checkOutbound(o.tag)))
-  } finally {
-    testingAll.value = false
-  }
-}
+// ---------------- delay check ----------------
+const {
+  checkResults,
+  testingAll,
+  check: checkOutbound,
+  checkAll: checkAllOutbounds,
+  delayRow,
+} = useDelayCheck(() => outbounds.value)
 
 // ---------------- card rows ----------------
 const cardRows = (item: any): EntityRow[] => [
@@ -147,20 +122,8 @@ const cardRows = (item: any): EntityRow[] => [
     v: Object.hasOwn(item, 'tls') ? t(item.tls?.enabled ? 'enable' : 'disable') : t('ui.none'),
     color: item.tls?.enabled ? 'var(--emerald)' : undefined,
   },
-  delayRow(item),
+  delayRow(item.tag),
 ]
-
-const delayRow = (item: any): EntityRow => {
-  const r = checkResults.value[item.tag]
-  if (r?.loading) return { k: t('out.delay'), v: '…', mono: true }
-  if (r && r.loading == false) {
-    if (r.success) {
-      return { k: t('out.delay'), v: (r.data?.Delay ?? 0) + ' ' + t('date.ms'), mono: true, color: 'var(--emerald)' }
-    }
-    return { k: t('out.delay'), v: r.errorMessage || t('failed'), color: 'var(--rose)' }
-  }
-  return { k: t('out.delay'), v: t('ui.none') }
-}
 
 // ---------------- drawers / modals ----------------
 const drawer = ref({ visible: false, id: 0, data: '' })
